order_service_app/order_service/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
mysqlclient
pymysql
gunicorn
python-dotenv
cryptography>=41.0
requests
django-extensions>=3.2
pika>=1.3.0
tenacity>=8.2.0
